+++ src/app/App.tsx
@@ -1,10 +1,10 @@
 import * as React from "react"
-import {ChakraProvider, theme,} from "@chakra-ui/react"
-import {CartProvider} from "./context/CartContext";
-import {Layout} from "./UI/Layout";
-import AppRouter from "./router/AppRouter";
-import {CategoryProvider} from "./context/CategoryContext";
-import {CustomerProvider} from "./context/CustomerContext";
+import { ChakraProvider, theme, } from "@chakra-ui/react"
+import { CartProvider } from "../context/CartContext";
+import { Layout } from "../UI/Layout";
+import AppRouter from "../router/AppRouter";
+import { CategoryProvider } from "../context/CategoryContext";
+import { CustomerProvider } from "../context/CustomerContext";
 
 export const App = () => {
     return (
