+++ src/index.tsx
@@ -1,19 +1,18 @@
-import {ColorModeScript} from "@chakra-ui/react"
 import React, {StrictMode} from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from "react-router-dom";
-import {App} from "./App"
-import './global.css';
+import {App} from "./app/App"
+import './app/styles/global.css';
 import "slick-carousel/slick/slick.css";
 import "slick-carousel/slick/slick-theme.css";
 import 'moment/locale/ru';
+import './config/i18n/i18n';
 
 const rootElement = document.getElementById('root');
 const root = ReactDOM.createRoot(rootElement as Element);
 
 root.render(
     <StrictMode>
-        <ColorModeScript/>
         <BrowserRouter>
             <App/>
         </BrowserRouter>
