+++ src/shared/assets/images/favicon.ico
