+++ src/shared/assets/images/logo.png
